Add image boxes and some functionality
