+++ CSE/CSE/Program.cs
@@ -17,7 +17,7 @@ namespace CSE
         {
             Application.EnableVisualStyles();
             Application.SetCompatibleTextRenderingDefault(false);
-            Application.Run(new Form1());
+            Application.Run(new frmMain());
         }
     }
 }
